fixing issues

+++ output/example.experiment.sol
@@ -6,9 +6,9 @@ contract Contract {
         uint64 a = 12345;
         uint64 b = 1e18;
 
-assert (a <= type(uint256).max / b); //slytherin 1
-        uint c = a * b;
-        
+assert (a <= type(uint64).max / b); //slytherin 1
+        uint256 c = a * b;
+
         return c;
     }
